refactor(db): move sqlc packages under dialects
