+++ pkg/cqrs/base_cqrs/README.md
@@ -14,6 +14,6 @@ Planned destinations:
 - `pkg/cqrs/manager` for the composite manager
 - `pkg/cqrs` for caller-facing history contracts and payload types
 
-The old Postgres normalization layer has been removed. The remaining generated
-query code still lives under `sqlc/` only until the dialect-local move in phase
-2.
+The old Postgres normalization layer has been removed. Generated query code now
+lives with the dialect adapters under `pkg/db/sqlite/sqlc` and
+`pkg/db/postgres/sqlc`.
