+++ pkg/db/sqlite/convert.go
@@ -6,7 +6,7 @@ import (
 	"strings"
 	"time"
 
-	sqlc "github.com/inngest/inngest/pkg/cqrs/base_cqrs/sqlc/sqlite"
+	sqlc "github.com/inngest/inngest/pkg/db/sqlite/sqlc"
 	"github.com/inngest/inngest/pkg/db"
 )
