+++ pkg/db/sqlite/sqlc/decorator.go
